+++ packages/react-renderer-demo/src/components/common/landing-page-logo.js
