+++ packages/react-renderer-demo/src/components/landing-page/landing-page-cards.js
@@ -5,7 +5,10 @@ import Paper from '@material-ui/core/Paper';
 import Typography from '@material-ui/core/Typography';
 import Divider from '@material-ui/core/Divider';
 
-import CodeEditor from '../common/component/code-editor';
+import dynamic from 'next/dynamic'
+const CodeEditor = dynamic(import('../code-editor'), {
+  ssr: false
+});
 
 const value = `import React from 'react';
 import FormRenderer, { componentTypes } from '@data-driven-forms/react-form-renderer';
